+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # *XDHq* *CHANGELOG*
 
+## 2018-12-18
+
+- adding token handling through plugin,
+- handling error message when problem on connection (added to handle error on token handling)
+
 ## 2018-12-07
 
 - switching to new protocol, to allow handling of error messages in future evolution,